predict int64 to int32

+++ GYAK/GYAK05/GYAK05.py
@@ -42,7 +42,7 @@ class KNNClassifier:
             distances= np.array(sorted(zip(distances,self.y_train)))
             label_pred = mode(distances[:self.k,1], keepdims=False).mode
             labels_pred.append(label_pred)
-            self.y_preds = np.array(labels_pred, dtype=np.int64)
+            self.y_preds = np.array(labels_pred, dtype=np.int32)
 
     def accuracy(self) -> float:
         true_positive = (self.y_test == self.y_preds).sum()
